etl/requirements.txt
# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Google Sheets
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Validation
cerberus==1.3.5

# Logging & Utilities
colorlog==6.8.0
tqdm==4.66.1

# Alternative data sources
openpyxl==3.1.2  # For Excel files
requests==2.31.0  # For API calls